refactor: ディレクトリ構成を修正
ファイル配置を分かりやすくするため
