+++ frontend/src/App.tsx
@@ -8,7 +8,7 @@ import DivisionsNew from "./pages/DivisionsNew";
 import SignUp from "./pages/SignUp";
 import Header from "./components/Header";
 import Home from "./pages/Home";
-import { AuthProvider } from "./components/AuthProvider";
+import { AuthProvider } from "./providers/AuthProvider";
 
 const App: React.FC = () => {
   return (
